choose mobilenet model and add crop postprocessing
