--- models/public/deblurgan-v2/deblurgan-v2-mobilenet/model.yml
@@ -1,74 +1,0 @@
-# Copyright (c) 2020 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-description: >-
-  DeblurGAN-v2 is a generative adversarial network (GAN) for single image motion deblurring.
-  This model is based on a relativistic conditional GAN with a double-scale discriminator.
-  Model used MobileNet as backbone and was trained on GoPro dataset.
-  For details about implementation of model,
-  check out the DeblurGAN-v2: Deblurring (Orders-of-Magnitude) Faster and Better <https://github.com/VITA-Group/DeblurGANv2>
-  repository.
-task_type: image_processing
-files:
-  - name: models/__init__.py
-    size: 0
-    sha256: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
-    source: https://raw.githubusercontent.com/VITA-Group/DeblurGANv2/master/models/__init__.py
-  - name: models/fpn_mobilenet.py
-    size: 5688
-    sha256: 1b24574ed94f6bb037b91208c5bcde9da243d2a175493ca942e448176c529efe
-    source: https://raw.githubusercontent.com/VITA-Group/DeblurGANv2/master/models/fpn_mobilenet.py
-  - name: models/mobilenet_v2.py
-    size: 4214
-    sha256: 16801e691c8bdaf16d57c00d3511fe1357d0a6fe63cf221f7c48dbc491383bc7
-    source: https://raw.githubusercontent.com/VITA-Group/DeblurGANv2/master/models/mobilenet_v2.py
-  - name: models/networks.py
-    size: 12779
-    sha256: 3857345de80c42dedc5162d3e82a173160041cf4e0b307fd7e29680a1115d097
-    source: https://raw.githubusercontent.com/VITA-Group/DeblurGANv2/master/models/networks.py
-  - name: ckpt/fpn_mobilenet.h5
-    size: 13505344
-    sha256: 6d1b6d9807da3ae8987021c063a5cb6ca0131e76388cdb5164404ee1b1bf522b
-    source:
-      $type: google_drive
-      id: 1JhnT4BBeKBBSLqTo6UsJ13HeBXevarrU
-postprocessing:
-  - $type: regex_replace
-    file: models/networks.py
-    pattern: 'from models\.(fpn_inception|fpn_inception_simple|unet_seresnext|fpn_densenet) '
-    replacement: '# \g<0>'
-  - $type: regex_replace
-    file: models/fpn_mobilenet.py
-    pattern: 'pretrained=True'
-    replacement: 'pretrained=False'
-conversion_to_onnx_args:
-  - --model-path=$config_dir
-  - --model-path=$dl_dir
-  - --model-name=DeblurV2
-  - --import-module=model
-  - --input-shape=1,3,736,1312
-  - --output-file=$conv_dir/deblurgan-v2.onnx
-  - --model-param=weights=r"$dl_dir/ckpt/fpn_mobilenet.h5"
-  - --model-param=model_name="fpn_mobilenet"
-  - --input-names=blur_image
-  - --output-names=deblur_image
-model_optimizer_args:
-  - --model_name=deblurgan-v2
-  - --reverse_input_channels
-  - --input_shape=[1,3,736,1312]
-  - --input=blur_image
-  - --output=deblur_image
-  - --input_model=$conv_dir/deblurgan-v2.onnx
-framework: pytorch
-license: https://raw.githubusercontent.com/CSAILVision/semantic-segmentation-pytorch/9aff40de31ee4b21f18514d31e5d6e4ba056924d/LICENSE
