+++ models/public/deblurgan-v2/model.yml
@@ -15,51 +15,43 @@
 description: >-
   DeblurGAN-v2 is a generative adversarial network (GAN) for single image motion deblurring.
   This model is based on a relativistic conditional GAN with a double-scale discriminator.
-  Model used Inception-ResNet-v2 as backbone and was trained on GoPro dataset.
-  For details about implementation of model,
-  check out the DeblurGAN-v2: Deblurring (Orders-of-Magnitude) Faster and Better <https://github.com/VITA-Group/DeblurGANv2>
-  repository.
+  For details about architecture of model, check out the paper <https://arxiv.org/abs/1908.03826>.
+  Model used Mobilenet as backbone and was trained on GoPro, DVD, NFS datasets. For
+  details about implementation of model, check out the DeblurGAN-v2: Deblurring (Orders-of-Magnitude)
+  Faster and Better <https://github.com/VITA-Group/DeblurGANv2> repository.
 task_type: image_processing
 files:
   - name: models/__init__.py
     size: 0
     sha256: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
     source: https://raw.githubusercontent.com/VITA-Group/DeblurGANv2/master/models/__init__.py
-  - name: models/fpn_inception.py
-    size: 6466
-    sha256: baf7f973519c30644e6ba0f95ef73376963991c34bfc533e32dcc94fe15e51a0
-    source: https://raw.githubusercontent.com/VITA-Group/DeblurGANv2/master/models/fpn_inception.py
-  - name: models/inceptionresnetv2.py
-    size: 12055
-    sha256: 850ef83c94bcbd2fd5c89ddbbe88afb4d61315b50b5715d4988d028fd1615d5d
-    source: https://raw.githubusercontent.com/Cadene/pretrained-models.pytorch/master/pretrainedmodels/models/inceptionresnetv2.py
+  - name: models/fpn_mobilenet.py
+    size: 5688
+    sha256: 1b24574ed94f6bb037b91208c5bcde9da243d2a175493ca942e448176c529efe
+    source: https://raw.githubusercontent.com/VITA-Group/DeblurGANv2/master/models/fpn_mobilenet.py
+  - name: models/mobilenet_v2.py
+    size: 4214
+    sha256: 16801e691c8bdaf16d57c00d3511fe1357d0a6fe63cf221f7c48dbc491383bc7
+    source: https://raw.githubusercontent.com/VITA-Group/DeblurGANv2/master/models/mobilenet_v2.py
   - name: models/networks.py
     size: 12779
     sha256: 3857345de80c42dedc5162d3e82a173160041cf4e0b307fd7e29680a1115d097
     source: https://raw.githubusercontent.com/VITA-Group/DeblurGANv2/master/models/networks.py
-  - name: ckpt/deblur.h5
-    size: 244480361
-    sha256: 0ab85a0419534bca6c8bee65eff3c163313e647b4577d493d679aa31169b0eea
+  - name: ckpt/fpn_mobilenet.h5
+    size: 13505344
+    sha256: 6d1b6d9807da3ae8987021c063a5cb6ca0131e76388cdb5164404ee1b1bf522b
     source:
       $type: google_drive
-      id: 1UXcsRVW-6KF23_TNzxw-xC0SzaMfXOaR
+      id: 1JhnT4BBeKBBSLqTo6UsJ13HeBXevarrU
 postprocessing:
   - $type: regex_replace
     file: models/networks.py
-    pattern: 'from models\.(fpn_mobilenet|fpn_inception_simple|unet_seresnext|fpn_densenet) '
+    pattern: 'from models\.(fpn_inception|fpn_inception_simple|unet_seresnext|fpn_densenet) '
     replacement: '# \g<0>'
   - $type: regex_replace
-    file: models/fpn_inception.py
-    pattern: 'from pretrainedmodels import inceptionresnetv2'
-    replacement: 'from .inceptionresnetv2 import inceptionresnetv2'
-  - $type: regex_replace
-    file: models/fpn_inception.py
-    pattern: "pretrained='imagenet'"
-    replacement: 'pretrained=None'
-  - $type: regex_replace
-    file: models/fpn_inception.py
-    pattern: 'from torchsummary import summary'
-    replacement: '# \g<0>'
+    file: models/fpn_mobilenet.py
+    pattern: 'pretrained=True'
+    replacement: 'pretrained=False'
 conversion_to_onnx_args:
   - --model-path=$config_dir
   - --model-path=$dl_dir
@@ -67,8 +59,8 @@ conversion_to_onnx_args:
   - --import-module=model
   - --input-shape=1,3,736,1312
   - --output-file=$conv_dir/deblurgan-v2.onnx
-  - --model-param=weights=r"$dl_dir/ckpt/deblur.h5"
-  - --model-param=model_name="fpn_inception"
+  - --model-param=weights=r"$dl_dir/ckpt/fpn_mobilenet.h5"
+  - --model-param=model_name="fpn_mobilenet"
   - --input-names=blur_image
   - --output-names=deblur_image
 model_optimizer_args:
